+++ attributes/authoritative.rb
@@ -17,26 +17,19 @@
 # limitations under the License.
 #
 
-default['pdns']['authoritative']['config_dir'] = '/etc/powerdns'
+default['pdns']['authoritative']['backends'] = %w( bind )
+default['pdns']['authoritative']['config']['launch'] = 'bind'
 
-# For Source based builds
-default['pdns']['authoritative']['source']['version'] = '3.4.7'
-default['pdns']['authoritative']['source']['url'] = "https://downloads.powerdns.com/releases/pdns-#{node['pdns']['authoritative']['source']['version']}.tar.bz2"
-default['pdns']['authoritative']['source']['path'] = '/opt'
-default['pdns']['authoritative']['source']['backends'] = %w( pipe )
+default['pdns']['authoritative']['config']['config_dir'] = '/etc/powerdns'
+default['pdns']['authoritative']['config']['setgid'] = 'pdns'
+default['pdns']['authoritative']['config']['setuid'] = 'pdns'
+default['pdns']['authoritative']['config']['version_string'] = 'powerdns'
 
-# For Packages based builds
-default['pdns']['authoritative']['package']['backends'] = %w( pipe )
-
-default['pdns']['authoritative']['config']['config_dir'] = node['pdns']['authoritative']['config_dir']
 default['pdns']['authoritative']['config']['allow_recursion'] = [ '127.0.0.1' ]
 default['pdns']['authoritative']['config']['daemon'] = true
 default['pdns']['authoritative']['config']['disable_axfr'] = true
 default['pdns']['authoritative']['config']['guardian'] = true
-default['pdns']['authoritative']['config']['setgid'] = node['pdns']['group']
-default['pdns']['authoritative']['config']['setuid'] = node['pdns']['user']
-default['pdns']['authoritative']['config']['version_string'] = 'powerdns'
 default['pdns']['authoritative']['config']['default_ttl'] = '3600'
+
 default['pdns']['authoritative']['config']['pipe_command'] = '/opt/pdns-3.4.4/modules/pipebackend/backend.pl'
 default['pdns']['authoritative']['config']['launch'] = 'pipe'
-