+++ attributes/recursor.rb
@@ -19,7 +19,7 @@
 
 default['pdns']['recursor']['config_dir'] = '/etc/powerdns'
 
-default['pdns']['recursor']['config']['config_dir'] = node['pdns']['recursor']['config_dir']
+default['pdns']['recursor']['config']['config_dir'] = '/etc/powerdns'
 default['pdns']['recursor']['config']['allow_from'] = [
   '127.0.0.0/8',
   '10.0.0.0/8',