+++ attributes/slave.rb
@@ -1,6 +1,6 @@
 #
 # Cookbook Name:: pdns
-# Attributes:: default
+# Attributes:: slave
 #
 # Copyright 2014, Aetrion, LLC.
 #
@@ -17,7 +17,15 @@
 # limitations under the License.
 #
 
-default['pdns']['build_method'] = 'package'
-default['pdns']['flavor'] = 'recursor'
-default['pdns']['user'] = 'pdns'
-default['pdns']['group'] = 'pdns'
+default['pdns']['slave']['backends'] = %w( bind )
+default['pdns']['slave']['config']['launch'] = 'bind'
+
+default['pdns']['slave']['config']['config_dir'] = '/etc/powerdns'
+default['pdns']['slave']['config']['setgid'] = 'pdns'
+default['pdns']['slave']['config']['setuid'] = 'pdns'
+default['pdns']['slave']['config']['version_string'] = 'powerdns'
+
+default['pdns']['slave']['config']['master'] = false
+default['pdns']['slave']['config']['slave'] = true
+default['pdns']['slave']['config']['slave_cycle_interval'] = '60'
+default['pdns']['slave']['config']['disable_axfr'] = true